Allow for using multiple GH personal tokens
One token only provides 5000 calls an hour, this limit is super low when wanting to scale more rapidly.  I want to refactor this more to allow for it to use x amount of tokens and not be as hacky, but for now this has the desired result.

Moved runner pagination to 100 from 10 as that was making additional uncessary calls.